requirements.txt
Pillow
python-sat